refactor: Add usage guide for SteamaCo Plugin + minor review change

+++ src/backend/database/migrations/tenant/2026_06_04_104602_change_amount_data_type_on_steama_transactions_table.php
@@ -7,7 +7,7 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up(): void {
         Schema::connection('tenant')->table('steama_transactions', function (Blueprint $table) {
-            $table->decimal('amount', 10, 2)->change();
+            $table->float('amount')->change();
         });
     }
